refactor: Move auth request DTOs to request subpackage
