+++ identity_data/src/commonMain/kotlin/net/thechance/mena/identity/data/dto/auth/request/RegisterRequestDto.kt
@@ -1,4 +1,4 @@
-package net.thechance.mena.identity.data.dto.auth
+package net.thechance.mena.identity.data.dto.auth.request
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
